Remove read_trials_from_remote_storage
